backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
psycopg[binary]==3.2.10
pgvector==0.2.5
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
